- initiate room catalog
