+++ lib/components/Reservation/index.js
@@ -11,6 +11,7 @@ import I18n from 'react-native-i18n';
 import moment from 'moment';
 import { get } from 'lodash/object';
 import { last, first, take } from 'lodash/array';
+import { isEmpty } from 'lodash/lang'
 import { filter, groupBy, orderBy } from 'lodash/collection';
 import keyBy from 'lodash/keyBy';
 import Icon from 'react-native-vector-icons/FontAwesome';
@@ -214,6 +215,36 @@ class Reservation extends Component {
     }
 
     const sortReservationData = orderBy(reservations, [reservation => get(reservation, 'bedName', '') ? get(reservation, 'bedName', '').toLowerCase() : ''], ['asc']);
+    const withoutFutureArr = filter(sortReservationData, function (obj) {
+      const today = moment().format('YYYY-MM-DD');
+      const formateCheckIn = moment(get(obj, 'mCheckin', null)).format('YYYY-MM-DD')
+      if (String(get(obj, 'guestStatus', null)).trim() === 'ARR') {
+        if (formateCheckIn === today) {
+          return true
+        } else {
+          
+          return false
+        }
+      } else {
+        return true
+      }
+    })
+
+    const withFutureArr = filter(sortReservationData, function (obj) {
+      const today = moment().format('YYYY-MM-DD');
+      const formateCheckIn = moment(get(obj, 'mCheckin', null)).format('YYYY-MM-DD')
+      if (String(get(obj, 'guestStatus', null)).trim() === 'ARR') {
+        if (formateCheckIn === today) {
+          return false
+        } else {
+          return true
+        }
+      } else {
+        return false
+      }
+    })
+
+    const validateReservationList = !isEmpty(withoutFutureArr) ? withoutFutureArr : withFutureArr
 
     return (
       <View style={[style]}>
@@ -222,9 +253,9 @@ class Reservation extends Component {
           <Text style={styles.title}>{I18n.t('attendant.components.reservation-component.reservation').toUpperCase()}</Text>
         }
 
-        {isHorizontal && sortReservationData.length > 1 ?
+        {isHorizontal && validateReservationList.length > 1 ?
           <ScrollView horizontal contentContainerStyle={{ height: 100, marginBottom: 10 }}>
-            {sortReservationData.map((reservation, index) =>
+            {validateReservationList.map((reservation, index) =>
               <Entry
                 key={index}
                 isActive={activeId === reservation.guest_id}
@@ -236,7 +267,7 @@ class Reservation extends Component {
           </ScrollView>
           :
           <View>
-            {sortReservationData.map((reservation, index) =>
+            {validateReservationList.map((reservation, index) =>
               <Entry
                 key={index}
                 isActive={activeId === reservation.guest_id}
