+++ lib/layouts/Gallery/index.js
@@ -1,193 +1,136 @@
-import React, { Component } from 'react';
+import React, { Component, createRef } from 'react';
 import {
-  View,
-  ScrollView,
-  Text,
-  Image,
-  ImageBackground,
-  Dimensions,
-  StyleSheet,
-  Modal
-} from 'react-native';
-
-import ListView from 'rc-mobile-base/lib/components/ListView';
-import Lightbox from 'react-native-lightbox';
-import Icon from 'react-native-vector-icons/FontAwesome';
-import Entypo from 'react-native-vector-icons/Entypo';
-import Pdf from 'react-native-pdf';
-import ImageViewer from 'react-native-image-zoom-viewer';
-
-import { connect } from 'react-redux';
-import { getRoomById, getCatalogByRoomId } from 'rc-mobile-base/lib/selectors/rooms';
-
-import {
-  Container,
-  ImageContainer,
-  ImageContent,
-  DescriptionContainer,
-  DescriptionText,
-  PdfContainer,
-  PdfContent,
-  FocusImageClose,
-  PdfFullContainer,
-  PdfExitContainer
+  styles,
+  SafeAreaContainer,
+  RootContainer,
+  PageContainer,
+  HeaderImageSliderContainer,
+  HeaderImages,
+  ImageDescriptionContainer,
+  DescriptionFlatList
 } from './styles';
 
+import { isEmpty, get } from 'lodash';
+import PagerView from 'react-native-pager-view';
 import {
-  lCenterCenter,
-  white,
-  slate
-} from 'rc-mobile-base/lib/styles';
+  PagerTabIndicator,
+  IndicatorViewPager,
+  PagerTitleIndicator,
+  PagerDotIndicator,
+} from '@shankarmorwal/rn-viewpager';
 
-const GalleryImage = ({ item, handler = () => null }) => (
-  <ImageContainer>
-    <ImageContent
-      onPress={handler}
-      >
-      <ImageBackground
-          style={styles.largeWideImage}
-          source={{ uri: item && item.image || '' }}
-        >
-        <View style={styles.tile}>
-          <Text style={styles.title}>{ item.comment }</Text>
-        </View>
-      </ImageBackground>
-    </ImageContent>
+import request from 'rc-mobile-base/lib/utils/request';
+import { connect } from 'react-redux';
+import { tokenSelector, hotelGroupKeySelector, apiSelector } from 'rc-mobile-base/lib/selectors/auth';
 
-    { item.description ?
-      <DescriptionContainer>
-        <DescriptionText>{ item.description }</DescriptionText>
-      </DescriptionContainer>
-      : null
-    }
-  </ImageContainer>
-)
+const ROOM_CATALOGS_API = `/RoomCatalog/GetListOfRoomCatalogImages`;
 
-const GalleryPdf = ({ item, handler = () => null }) => (
-  <PdfContainer>
-    <PdfContent onPress={handler}>
-      <Icon name={'file-pdf-o'} color={slate.color} size={64} />
-      
-      <View style={styles.tile}>
-        <Text style={styles.title}>{ item.comment }</Text>
-      </View>
-    </PdfContent>
-
-    { item.description ?
-      <DescriptionContainer>
-        <DescriptionText>{ item.description }</DescriptionText>
-      </DescriptionContainer>
-      : null
-    }
-  </PdfContainer>
-)
+// function* fetchRoomCatalogs() {
+//   return yield call(authRequest, ROOM_CATALOGS_API);
+// }
 
 class GalleryLayout extends Component {
   constructor(props) {
     super(props);
 
+    this.pagerRef = createRef()
+
     this.state = {
-      activeImage: null,
-      activePdf: null
+      catalogHolder: []
     }
+
   }
 
-  static navigationOptions = {
-    title: 'Gallery',
+  componentDidMount() {
+    this.getRoomCatalog()
+
+  }
+
+  getRoomCatalog = async () => {
+    const { token, hotel_group_key, baseUrl } = this.props;
+    const roomId = this.props.navigation.state.params.roomId
+
+    const headers = {
+      Authorization: `Bearer ${token}`,
+      hotel_group_key: hotel_group_key,
+      'Content-Type': 'application/json'
+    };
+
+    const url = `${baseUrl}${ROOM_CATALOGS_API}`
+    const options = {
+      method: 'POST',
+      body: JSON.stringify({
+        roomId: roomId
+      })
+    }
+
+    const catalogList = await request(url, {
+      ...options,
+      headers,
+    });
+
+    if (!isEmpty(catalogList)) {
+      console.log('--- catalogList ---');
+      console.log(catalogList)
+
+      this.setState({ catalogHolder: catalogList })
+    }
+
+  }
+
+  renderViewPagerPage = (data, index) => {
+    const imageUrl = get(data, 'annotatedImageUrl', null);
+    console.log('--- imageUrl ---', imageUrl)
+    return (
+      <PageContainer key={data}>
+        <HeaderImages source={{ uri: imageUrl }} />
+      </PageContainer>
+    );
   };
 
-  componentWillMount() {
-    const { room: { name }} = this.props;
+  stepFromPagination = async (position) => {
+    console.log('--- stepFromPagination ---');
+    console.log(position)
   }
-
-  _handlePress(item) {
-    const { activeItem } = this.state;
-    this.setState({ activeItem: activeItem === item ? null : item });
-  }
-
-  _setActivePdf = (pdf) => this.setState({ activePdf: pdf, activeImage: null })
-  _setActiveImage = (image) => this.setState({ activeImage: image, activePdf: null })
 
   render() {
-    const { activePdf, activeImage } = this.state;
-    const { catalog } = this.props;
-
-    if (activeImage) {
-      return (
-        <Modal visible={true} transparent={true}>
-          <ImageViewer
-            imageUrls={[{ url: activeImage },]}
-            enableSwipeDown
-            onSwipeDown={() => this.setState({ activeImage: null })}
-            renderHeader={() => <FocusImageClose handler={() => this.setState({ activeImage: null })} />}
-            renderInd
-            />
-        </Modal>
-      )
-    }
-    console.log(activePdf)
-
-    if (activePdf) {
-      return (
-        <PdfFullContainer>
-          <Pdf
-            style={{ flex: 1 }}
-            source={{ uri: activePdf }}
-            />
-          <PdfExitContainer onPress={() => this._setActivePdf(null)}>
-            <Entypo name="cross" size={24} color="black" />
-          </PdfExitContainer>
-        </PdfFullContainer>
-      )
-    }
+    const { catalogHolder } = this.state;
+    const pageCount = get(catalogHolder, 'length');
 
     return (
-      <Container>
-        { catalog.map(item =>
-          item.isPdf ?
-          <GalleryPdf
-            key={item._id}
-            item={item}
-            handler={() => this._setActivePdf(item.image)}
-            />
-          :
-          <GalleryImage
-            key={item._id}
-            item={item}
-            handler={() => this._setActiveImage(item.image)}
-            />
-        )}
-      </Container>
+      <SafeAreaContainer>
+        <RootContainer>
+          {/* <PagerView
+            style={{ flexGrow: 1 }}
+            initialPage={0}
+            ref={this.pagerRef}
+          // onPageSelected={(e) => this.stepFromPagination(e.nativeEvent.position)}
+          >
+            {catalogHolder.map((page, index) => this.renderViewPagerPage(page, index))}
+          </PagerView> */}
+          <HeaderImageSliderContainer>
+            <IndicatorViewPager
+              style={styles.pagerStyle}
+              indicator={
+                <PagerDotIndicator pageCount={pageCount} />
+              }
+              onPageScrollStateChanged={(e) => this.stepFromPagination(e)}
+              >
+              {catalogHolder.map((page, index) => this.renderViewPagerPage(page, index))}
+            </IndicatorViewPager>
+          </HeaderImageSliderContainer>
+
+          <ImageDescriptionContainer>
+              {/* <DescriptionFlatList 
+                data={}
+              /> */}
+          </ImageDescriptionContainer>
+
+        </RootContainer>
+      </SafeAreaContainer>
     )
   }
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    paddingTop: 54,
-    paddingLeft: 0,
-    paddingRight: 0,
-    paddingBottom: 10,
-    backgroundColor: '#F7F7F7',
-  },
-  largeWideImage: {
-    height: 240
-  },
-  tile: {
-    position: 'absolute',
-    left: 0,
-    right: 0,
-    bottom: 0,
-    height: 40,
-    backgroundColor: 'rgba(0,0,0,.5)',
-    ...lCenterCenter
-  },
-  title: {
-    ...white.text,
-    fontSize: 17
-  }
-});
 
 const mapStateToProps = (state, props) => {
   const roomId = props.navigation.state.params.roomId
@@ -196,8 +139,9 @@ const mapStateToProps = (state, props) => {
   // console.log(isAttendant, isRunner);
 
   return {
-    room: getRoomById(roomId)(state),
-    catalog: getCatalogByRoomId(roomId)(state)
+    token: tokenSelector(state),
+    hotel_group_key: hotelGroupKeySelector(state),
+    baseUrl: apiSelector(state)
   }
 };
