+++ lib/sagas/rooms.js
@@ -42,6 +42,7 @@ export default function ({ apiUrl, userType }) {
   const CALENDAR_API = `/Reservation/GetListOfReservations`;
   const ROOM_NOTES_API = `/Room/GetListOfNotes`;
   const CATALOGS_API = `/catalog_by_hotel`;
+  const ROOM_CATALOGS_API = `/RoomCatalog/GetListOfRoomCatalogImages`;
   const TASKS_API = `/Task/GetListOfTasks`;
   const HISTORY_API = `/hotel_history`;
   const GUEST_BOOK_API = `/guest_book`;
@@ -549,6 +550,11 @@ export default function ({ apiUrl, userType }) {
     yield throttle(3000, RoomsTypes.ROOM_NOTES_FETCH, fetchRoomNotesFlow);
   }
 
+  // Room Catalogs
+  function* fetchRoomCatalogs() {
+    return yield call(authRequest, ROOM_CATALOGS_API);
+  }
+
   // Hotel Catalogs
   function* fetchCatalogs() {
     return yield call(authRequest, CATALOGS_API);
@@ -882,6 +888,7 @@ export default function ({ apiUrl, userType }) {
       fetchCalendarFlow,
       fetchRoomNotes,
       fetchRoomNotesFlow,
+      fetchRoomCatalogs,
       fetchCatalogs,
       fetchCatalogsFlow,
       fetchLostItems,
